just a little clean up, moved some interfaces to main interface file and removed service worker
