+++ src/interfaces.d.ts
@@ -41,3 +41,35 @@ export interface IStore {
   state: IState;
   dispatch?: React.Dispatch<IAction>;
 }
+
+export interface Iqvar {
+  name: string;
+  xpath: string;
+}
+export interface ISubstitution {
+  name: string;
+  term: string;
+}
+
+export interface IFormulaHit {
+  id: number;
+  local_id: number;
+  /** the document the hit comes fromt*/
+  segment: string;
+  /** the title from metadata */
+  title?: string;
+  /** the url to the fromula */
+  url?: string;
+  /** the mathml for the complete formula */
+  source: string;
+  /** the mathml for the subterm that was hit with the searchquery  */
+  subterm?: string;
+  /** the xpath to subterm in source */
+  xpath: string;
+  /** Array of the terms of subsitions */
+  substituitons?: ISubstitutions;
+  /** Query variables xpath */
+  queryvariablesxpath: Array<Iqvar>;
+  /** the text from the document */
+  text: string;
+}
